fix: Create dedicated utils module
